PARQUET-305: Update logging to SLF4J.
This removes the Log implementation based on java.util.logging and
replaces it with SLF4J. The compiler removal of debug log messages still
works because Log.DEBUG and similar final constants are unchanged.

This commit adds slf4j-simple as the test logger implementation.
Configuration for slf4j-simple is in the root pom. Two modules can't use
slf4j-simple, parquet-pig and parquet-thrift, and use slf4j-log4j12
instead because pig depends on log4j and tests die without it.

Author: Ryan Blue <blue@apache.org>

Closes #290 from rdblue/PARQUET-305-update-logging and squashes the following commits:

89257e8 [Ryan Blue] PARQUET-305: Remove deprecation annotations on Log.
9f9b99a [Ryan Blue] PARQUET-305: Update logging to SLF4J.