PARQUET-422: Fix a potential bug in MessageTypeParser where we ignore…
… and overwrite the initial value of a method parameter

In org.apache.parquet.schema.MessageTypeParser, for addGroupType() and addPrimitiveType(), the initial value of this parameter t is ignored, and t is overwritten here.

This often indicates a mistaken belief that the write to the parameter will be conveyed back to the caller.

This is a bug found by FindBugs™.

Author: proflin <proflin.me@gmail.com>

Closes #308 from proflin/PARQUET-422 and squashes the following commits:

df1f908 [proflin] PARQUET-422: Fix a potential bug in MessageTypeParser where we ignore and overwrite the initial value of a method parameter